Move Config and ACL to the top-level module (#20)
I would like to programatically generate relays configurations, however this
is slightly more difficult because, despite being exported types, they are in
the "main" package.

This PR moves them to the top-level directory so that they can be imported from
other modules.

I also took the liberty of documenting things.